refactor: extract shared MediaViewerShell for image and video viewers
Deduplicate the overlay bar (sender info, download, close), auto-hide
timer, and fullscreen dialog logic into a shared MediaViewerShell
widget. Both full_image_view and full_video_view now delegate to it.
